requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-filter==21.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0
unidecode==1.3.4
pillow==9.1.1